+++ vitest.config.js
@@ -18,6 +18,7 @@ export default defineConfig(() => {
     ],
     test: {
       environment: 'node',
+      include: ['tests/**/*.test.ts'],
       hookTimeout: 30_000,
       testTimeout: 30_000,
     },